requirements.txt
asgiref==3.4.1
Django==3.2.13
django-filter==21.1
djangorestframework==3.13.1
mysqlclient==2.1.0
Pillow==8.4.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
